+++ src/state/slices/swappersSlice/swappersSlice.ts
@@ -1,27 +1,13 @@
 import type { PayloadAction } from '@reduxjs/toolkit'
 import { createSlice } from '@reduxjs/toolkit'
-import { createApi } from '@reduxjs/toolkit/query/react'
-import type { AccountId, AssetId } from '@shapeshiftoss/caip'
+import type { AccountId } from '@shapeshiftoss/caip'
 import { ethAssetId, foxAssetId } from '@shapeshiftoss/caip'
-import type { Result } from '@sniptt/monads/build'
 import type { Asset } from 'lib/asset-service'
 import { localAssetData } from 'lib/asset-service'
 import { bnOrZero } from 'lib/bignumber/bignumber'
-import type {
-  GetEvmTradeQuoteInput,
-  GetTradeQuoteInput,
-  SwapErrorRight,
-  SwapperName,
-} from 'lib/swapper/api'
-import { getLifiTradeQuote } from 'lib/swapper/swappers/LifiSwapper/getTradeQuote/getTradeQuote'
-import type { LifiTradeQuote } from 'lib/swapper/swappers/LifiSwapper/utils/types'
-import { getThorTradeQuote } from 'lib/swapper/swappers/ThorchainSwapper/getThorTradeQuote/getTradeQuote'
-import { BASE_RTK_CREATE_API_CONFIG } from 'state/apis/const'
-import type { ReduxState } from 'state/reducer'
+import type { SwapperName } from 'lib/swapper/api'
 
 import { defaultAsset } from '../assetsSlice/assetsSlice'
-import { selectAssets, selectFeeAssetById } from '../assetsSlice/selectors'
-import { selectCryptoMarketData } from '../marketDataSlice/selectors'
 
 export enum StepTransactionStatus {
   Identified = 'identified', // The step is identified as being part of the execution flow
@@ -78,70 +64,3 @@ export const swappers = createSlice({
     },
   },
 })
-
-export const swappersApi = createApi({
-  ...BASE_RTK_CREATE_API_CONFIG,
-  reducerPath: 'swappersApi',
-  endpoints: build => ({
-    getLifiTradeQuote: build.query<
-      Result<LifiTradeQuote<false>, SwapErrorRight>,
-      GetTradeQuoteInput
-    >({
-      queryFn: async (getTradeQuoteInput: GetEvmTradeQuoteInput, { getState }) => {
-        const state: ReduxState = getState() as ReduxState
-        const assets: Partial<Record<AssetId, Asset>> = selectAssets(state)
-        // selectCryptoMarketData is used because prives are always in USD
-        const marketDataById = selectCryptoMarketData(state)
-        const sellAssetMarketData = marketDataById[getTradeQuoteInput.sellAsset.assetId]
-        if (!sellAssetMarketData)
-          return {
-            error: `no market data available for assetId ${getTradeQuoteInput.sellAsset.assetId}`,
-          }
-        const maybeQuote = await getLifiTradeQuote(
-          getTradeQuoteInput,
-          assets,
-          sellAssetMarketData.price,
-        )
-        return { data: maybeQuote }
-      },
-    }),
-    getThorTradeQuote: build.query<
-      Result<LifiTradeQuote<false>, SwapErrorRight>,
-      GetTradeQuoteInput
-    >({
-      queryFn: async (getTradeQuoteInput: GetTradeQuoteInput, { getState }) => {
-        const state: ReduxState = getState() as ReduxState
-        // selectCryptoMarketData is used because prives are always in USD
-        const marketDataById = selectCryptoMarketData(state)
-        const feeAsset = selectFeeAssetById(state, getTradeQuoteInput.sellAsset.assetId)
-        const sellAssetMarketData = marketDataById[getTradeQuoteInput.sellAsset.assetId]
-        const buyAssetMarketData = marketDataById[getTradeQuoteInput.buyAsset.assetId]
-        const feeAssetMarketData = feeAsset ? marketDataById[feeAsset.assetId] : undefined
-
-        if (!sellAssetMarketData)
-          return {
-            error: `no market data available for assetId ${getTradeQuoteInput.sellAsset.assetId}`,
-          }
-
-        if (!buyAssetMarketData)
-          return {
-            error: `no market data available for assetId ${getTradeQuoteInput.buyAsset.assetId}`,
-          }
-
-        if (!feeAssetMarketData)
-          return {
-            error: `no market data available for assetId ${feeAsset?.assetId}`,
-          }
-
-        const maybeQuote = await getThorTradeQuote(getTradeQuoteInput, {
-          sellAssetUsdRate: sellAssetMarketData.price,
-          buyAssetUsdRate: buyAssetMarketData.price,
-          feeAssetUsdRate: feeAssetMarketData.price,
-        })
-        return { data: maybeQuote }
-      },
-    }),
-  }),
-})
-
-export const { useGetLifiTradeQuoteQuery, useGetThorTradeQuoteQuery } = swappersApi
